Re-initialize RSpec
It will be helpful to use RSpec to check if our database is working
correctly, so here we begin getting that back to the default settings.
Running rspec --init generates a file that will configure RSpec to run
with the standard options. The verses have been brought out into a Ruby
constant for the time being.