Add DatabaseCleaner and configure
In the User model spec, I was manually deleting all the Users before and
after every test to ensure that we had a consistent state. This will
quickly become unmanageable as we add models and associations.
Here we add DatabaseCleaner to reset the test database for us. The
configuration is the default suggested by DatabaseCleaner's docs, which
leverages a database transaction to truncate the data-containing tables
quickly before the test suite and around each example.