Add User model timestamp tests
This might get squashed into an earlier commit but depending on the diff
and my goldfish brain I might not. This might also be a good candidate
for a shared example included on all model specs by default so we don't
ever have to look at this test again. Or we can delete it, it's dumb and
not likely to regress.