Add FactoryBot and configure
FactoryBot makes creating records in test really easy. Here we add it
and configure it for use in the existing User model specs and we can
remove a little bit of boilerplate test code. The email sequence is
tested by creating a list of 2 Users, we expect the number `n` to
increment, which makes the email unique on every User record created
this way.

User records do not have passwords, so at this point we have reached
parity at the model level and can delete the old User model specs.
Passwords and the resetting of those are no longer a concern of the
User model. We will offload that to our identity provider instead of
having to hand-roll that ourselves.