Add Commitments table
Here we add the new database table for Commitments. This represents a
year-long (by conventional) period of time where a user will commit to
doing activities. There's some basic validation for a start_date and
end_date in chronological order, and it provides some sensible defaults
so that a new Commitment starts "today" and ends after 1 year. We can
also create them with custom dates, but there's currently no mechanism
for setting the end_date based on the start_date. YAGNI.

This does introduce yet another testing dependency, ActiveSupport.
This includes the TimeHelpers that allow us to freeze time or travel to
a certain point in time to test some of the date-related logic.