Add Residents
Currently, Residents are loaded from a third party in bulk, so we'll
track the time that they were entered into the system. This allows us
to preserve historical data without relying on the created_at timestamp,
which really refers to the record itself, not the batch that the name
was in.
Because the names are coming from a third party, I don't trust the
cleanliness of the data, so names are validated case-insensitively.
This way we don't end up with repeated names with casing permutations:
e.g. Pete, pEte, peTe, and petE will all be considered the same name.