Add Prayer model
This commit adds a Prayer model to track prayers made by the user.
A prayer must have a User and the Resident that the user has prayed for.
A prayer may optionally have a Route if the prayer was made while the
user was on a route. The prayer also has a timestamp to record when the
prayer was made.