Add ApiSpecHelpers to parse responses
Adds a spec helper to easily parse the JSON representation of the
records we expect to be returned from the API. Because Timestamps are
returned in the default format using 3 digit precision in iso8601, the
spec helper does the translation for us. Now we don't need to add a
customserializer or a custom matcher for every spec example.