Add journeys
This commit adds the Journey model. Journeys are different goals
that a user can set for themselves. When a goal is set a new
Commitment is created. For now each Journey is to be completed in
one year, and the distances of each Journey depicts the distance
of a TN landmark that has been uploaded by an admin.