Add .rubocop.yml
I use rubocop locally, and while it's not enabled on this project yet,
it might be nice to move that direction sooner rather than later. Adding
this file to source control lets us use the same config and get the
same results. It would be great if we enabled rubocop to run at some
point before we merge a PR.

It ignores the block length because sometimes specs get long, ignores
the generated schema.rb file because it's not written by us, and
DEFINITELY ignores the documentation requirement because that's
annoying and you should be able to just read the code. We write good
commits so I don't think it's super necessary, but I'm open to a
discussion about it if you think documentation comments would be helpful.