Add route to POST Prayers
This route creates a new Prayer for the User and
includes the ID of the next Resident in the list
to minimize the number of requests we need to make.

There is a new validation on Prayers which ensures
that we aren't trying to save one without a
Resident and raising a database-level error,
resulting in a 500 status. This is a client error,
so we need to return a 4xx status.