Add verse_of_the_day method to Verse model
This commit adds a method to the Verse model that returns the verse
of the day. At the moment I have only been given 100 verses by the
client. I am handling this by dividing the day of the year by 100 and
using the remainder to select the verse of the day. This is not ideal
as it will repeat the verses every 100 days. Eventually when I have all
366 verses this will need to be changed.