Add Authentication middleware
This middleware sits in front of every endpoint and will retreive the
User based on the email header, which allows us to remove all the
duplicated code that used to do this.
The entire User object is stored in the request.env for access within
the action.
For whatever reason, the app.rb file is in scope when Factorybot runs,
so I've had to give the convenience method for the user a longer name
to avoid strange errors during test setup.

This also changes the header to something more realistic. When the
header originated from a real client, it's going to be all caps and
prefixed with `HTTP_`.