Add support for JWTs
We'll allow access to the API using a Bearer token, which will originate
from Amazon Cognito. This will accompany requests as a header,
effectively replacing the P4L-EMAIL header in the authentication
middleware. Verification of the JWTs is handled by the JsonWebToken
class, which relies on the JWKClient to fetch and cache the JWKs for
decoding an incoming JWT.

To simulate the token we'll get from Cognito, the
AuthenticationSpecHelpers now generate a JWT from an RSA key and export
the JWK to a MockJWKClient.

In tests, we always want to be using the MockJWKClient rather than the
real one, so by using `included` in the AuthenticationSpecHelpers we can
automatically stub this out for all tests that include it.