Simplify User and Verse seed task specs
Rake tasks do not play nicely with DatabaseCleaner for some reason.
Since the logic is contained in a class, it tends to be much easier to
test, so rather than invoke the Rake task from within the test, we can
just call the class directly and save ourselves the headache.

When the tests are run, the output from the db:seed:verses task prints
to the console, which we'd generally want to avoid. Wrapping the
execution of the task with the `output` matcher will do this. Along
the way, we can clean up this test to not share state between examples.

We can also use the bang version of find_or_create_by to raise
validation errors without an explicit `save!`.