Add sub to User
The `sub` is the primary key for the User in the Identity Provider
(IDP), which right now is only Amazon Cognito. This will uniquely
identify every User and we'll use this to allow access to that User's
data in our API. This effectively replaces the User email which we
temporarily used to look up a User.