Add JWKClient using HTTParty
Here we introduce the JWKClient, which also serves as our cache for JWKs
fetched from Cognito.
This is a Singleton so that we don't have to fetch the JWKs multiple
times, which would cause the app to respond slowly on startup until
every thread had populated its own cache.
Since this class makes outgoing HTTP requests, Webmock is added to
easily stub out these requests in tests.
When stubbing requests that you refer to later, make sure to use `let!`
to ensure that the request has been stubbed out before the request is
made.