Refactor data_object_for method
This commit refactors the data_object_for method in the
api_spec_helpers.rb file. The method now checks the entity to
determine whether it is an array. If it is an array, the method calls
the data_object_for_multiple method. If it is not an array, the method
operates as it did before.