Parse a users prayer list (pdf) and create new residents
This commit creates a new route /load_residents that will parse a pdf
file and create new residents based on the contents of the pdf. The
pdf file is expected to be a list of names and addresses. I have to
do some odd logic to account for empty lines or lines that do not
contain prayer names. I'm pretty sure that I am testing the incorrect
thing here because the pdf is attached as a param. The code does sort
the pdf and create the new residents. I just need to change how that
pdf is passed with the request. I'm also having an issue getting my
front end to make this call without getting a 401 error. I've been
using the same headers that I use for the other requests, so kinda
weird.