Create PdfParser service and specs
Creates a new service class to parse the PDF file and create the
User's Residents. The tests are a bit bare boned at the moment.
Would love to know what else I should test. I have also refactored
the the pdf parser to only eliminate any lines that contain a keyword
or are empty instead of taking lines in relation to the text in the
other lines. The only thing is I had to eliminate addresses based off
of whether or not there was a number in the line. I looked it up online
and some people say it is possible to have an address without a number
but I guess that is an edge case. Worst case someone will pray for an
address instead of a person.