Adds tests and functionality for POST /user/routes
Currently testing to ensure that an unauthorized user is unable to
create a route. Also testing to make sure that an authorized user
can create a new route and receive a Route object in return.

Also wrote some very simple model tests to ensure that the
current_commitment method is working as expected.