Makes first and last name optional for user
This changes the User model to allow for users to not have a first
or last name. This makes it more simple for new user's who are signing
up. We can add an info section to the settings screen where they can
add their first and last name if they want to. We could also add a pop
up when they first sign up.