Adds id_token to authentication spec helpers
This commit adds the id_token to the authentication spec helpers
so that the id_token can be required in the authentication middleware
without causing the tests to fail. The tests have also been changed to
test a new user as well as a user who's email address has changed. The
authentication code has also changed to take in only the JWT and not
the entire idToken object.

There are currently 2 failing tests in the route_spec and I can't seem
to figure out why. It doesn't look like the headers are coming through
with the request at all. I will definitely need to fix this before
merging.